LPS-134289 Rename files and remove .es extension
