+++ modules/apps/journal/journal-web/src/main/resources/META-INF/resources/js/JournalExportTranslation.js
