+++ modules/apps/translation/translation-web/package.json
@@ -13,7 +13,7 @@
 		"prop-types": "15.7.2",
 		"react": "16.12.0"
 	},
-	"main": "index.es.js",
+	"main": "index.js",
 	"name": "translation-web",
 	"scripts": {
 		"build": "liferay-npm-scripts build",
