+++ modules/apps/translation/translation-web/src/main/resources/META-INF/resources/js/export_translation/ExportTranslationContext.js
