+++ modules/apps/translation/translation-web/src/main/resources/META-INF/resources/js/export_translation/ExportTranslationModal.js
@@ -18,7 +18,7 @@ import ClayModal from '@clayui/modal';
 import PropTypes from 'prop-types';
 import React, {useContext, useState} from 'react';
 
-import ExportTranslationContext from './ExportTranslationContext.es';
+import ExportTranslationContext from './ExportTranslationContext';
 
 const noop = () => {};
